requirements.txt
Django==4.0.1
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-filter==23.1
pillow==9.5.0
drf-nested-routers==0.93.4
pytest-django==4.5.2
django-cors-headers==3.14.0